Register connection error metatype sooner
Move the EvernoteConnection::ErrorCode registration to the EvernoteConnection
class, so that it happens before the NoteStore and UserStore classes have a
chance to use it.
The UserStore class would otherwise emit an error when trying to get the
username, because the EvernoteConnection::ErrorCode type is not registered yet
at that point.