Make CroppedImageSizer async
QImageReader can block waiting on disk I/O so let's not have it on the main thread. 

CroppedImageSizerAsyncWorker is using a QFuture via QtConcurrent::run to get a thread and do the work. CroppedImageSizerAsyncWorker is created by CroppedImageSizer but deletes itself, either when the work has been finished or when it's aborted by CroppedImageSizer. This is why it's held with a QPointer in CroppedImageSizer.
Approved by: Michał Sawicz, Michael Zanetti