Don't block handling power events on loading the greeter's qml and the background image.
During power button press, we call greeter.showNow(), which was taking so long to return that we couldn't handle the power button release for a few seconds. Which meant that the shutdown dialog thought it would be a good time to appear.
                  
The easiest fix is to delay calling showNow until after we finish handling the event.
Approved by: Daniel d'Andrada, PS Jenkins bot