Don't force a surface to lose focus when its parent is updated.
This workarround prevents the closure of a trusted prompt session from causing the underlying surface to lose focus and then gain it right back. This brief loss of focus causes Qt.application.active to momentarily be set to 'false'. Currently, this is just a workarround and hopefuly by still setting focus to 'true' after the surface is added to the scene, active focus will still be updated correctly.